Convert OLMoE inference to f16 with mixed-precision stability
- All tensor data paths use f16 (params, activations, KV cache)
- Mixed precision where needed: f32 accumulation in RMS norm (overflow),
  f32 softmax in MoE router (exp overflow), f32 RoPE frequency computation
- iree_linalg_ext.attention keeps f32 internal accumulation body
- Scalar hyperparams (eps, rope_freq_base, scale) remain f32
- Add streaming token output and SIGINT handling to chat script
- Prefill: ~2.3s/token (was ~4.1s f32), Decode: ~2.8s/token on Orin MAXN

+++ components/kvcache/kvcache.mlir
@@ -32,7 +32,7 @@
 //   scatter_decode(cache, layer, new_k, new_v, block_tables, positions) -> cache
 //   scatter_prefill(cache, layer, new_k, new_v, block_tables, start_positions, block_size) -> cache
 
-!elem_t = f32
+!elem_t = f16
 
 module @kvcache_components {
